Fix bug in year_month sitemap generation
When the current month matches, we need to check that the year matches too, otherwise we set max days for all of the current months for past years as well. So triggering a sitemap generation on the 2nd of January would result in all January sitemaps only showing entries up to the 2nd of the month.