Allow non-cron-based sitemap builders
To prep the plugin for WP.com, let's abstract the cron-based functionality so that it can be disabled and overriden by a jobs-based sitemap builder.

Also, disable the admin pages when not using cron, because they're a bit rough around the edges.

We'll clean these up later and bring them back for everyone.