Renaming Create Sitemaps page title to be more consistant and removing the Sitemaps CPT admin page.
Renamed the "Create Sitemaps" page to "Sitemaps" and changed the header on
the page to match the menu name.

Removed the Sitemap custom post type admin page because it had no useful
purpose. See #53 for making this page useful.

Also renamed the sitemap_options() function to
render_sitemap_options_page() so that it was more self descriptive.

In response to comments from @mjangda here:
https://github.com/Mobius5150/msm-sitemap/commit/de3de21d396615001cb8875b2e01735736391207