Updating behaviour for halting sitemap generation.
Update the 'Halt Sitemap Generation' button so that it is only enabled when sitemap generation is actually in progress.

Added action handling text to the 'halt-generation' action so that the return message accurately reflects whether sitemap generation was halted or was already halted.

Added a status text to reflect when sitemap generation is being halted. This is because previously when a user clicked the halt sitemap generation button sitemap generation would be eventually stopped during the cron job, but would not be halted immediately. Adding the status text shows the user that the next time the cron job is processed sitemap generation will be halted -- giving them feedback after pushing the button.

In response to comments received on 864b296.