Cleaning up output formatting and page generation.
Added a space butween the 'Current Position' or 'Restart Position' text
and the following data.

Stats header is now an h3.

Buttons are only shown when they should be used. Ie: Generate and reset buttons are
only shown when sitemap generation is not running and the halt button is
only shown when the sitemap generation is running.

Buttons now use core styles rather than default buttons.