Fixing a status message bug on Sitemaps page.
Fixed a status message bug on the sitemaps page where if the user halted
generation the status text would continue to say halting after it was
restarted.

Also tweaked the behaviour of reset_sitemap_data() as this action is only
doable when sitemap processing is stopped - so it should not try and stop
sitemap processing.