Removing Metro_Sitemap dependance on MSM_Sitemap_Builder_Cron.
Removed the direct dependance from the Metro_Sitemap class on the
MSM_Sitemap_Builder_Cron class. This is because MSM Sitemaps are typically
built using WP CLI commands because the cron aproach does not scale well
when a large number of sitemaps need to be generated.

Added two filters to render_sitemap_options_page(): `msm_sitemap_actions`
and `msm_sitemap_create_status`. The first filter allows plugins to
register their actions to be shown on the admin page. The second allows
plugins to customize the status text shown to the end user.

When an action needs to be performed, the action
`msm_sitemap_action-$action` is called, where $action is a slug
provided by the plugin that created the action. If a plugin would like to
show a message to the user about the status or completion of that action
it can call Metro_Sitemap::show_action_message() with the action message
and the severity level of the message.

See the feedback received on #54.