Moving per-sitemap url counts to post metas.
Previously the total number of indexed URLs for each sitemap was stored in
a single option `msm_sitemap_indexed_url_count`, which ran the risk of
becoming much to large on sites with many sitemaps.

This change moves the count of urls in each individual sitemap to a
`msm_indexed_url_count` post meta for that sitemap.
`msm_sitemap_indexed_url_count` now hold the total number of indexed URLs.

Updated msm-sitemap.php and wp-cli.php to reflect these changes.

One side effect of this is that it is no longer possible to retrieve the
count for every sitemap on the site using `get_indexed_url_count()` as
that would require a large number of queries. Instead, this function is
now passed the year, month and day to retrieve.

Closes #60.