Updating UTs for the new sitemaps stats.
The method `get_indexed_url_count` changed in 8c1d7ab. Updating stats to
reflect this.

Added a helper function to `MSM_SiteMap_Test` that checks that the stats
are correct for all posts created using `create_dummy_posts()`. This
removes duplicate code for stat checking.

Note that the `test_site_stats_for_deleted_post()` test is currently
failing because sitemaps are not being removed when all posts from that
day are deleted. This test is failing correctly and the issue should be
fixed in another commit.