Do not autoload create in progress option.
The "msm_sitemap_create_in_progress" is critical to fully stop the
progress of the sitemap generation. It is stored as an autoloading
option, which is unfortunately susceptible to bizarre race conditions
(https://core.trac.wordpress.org/ticket/31245). Deleting this option can
be problematic because it gets stuck in the cache, thereby causing
problems with resetting the state.

This commit fixes the issue by explicitly setting the option to not be
autoloaded