Switch to raw SQL from WP_Date_Query (#93)
Using a direct comparison against post_date is significantly faster than
the date casting queries generated by WP_Date_Query. The former does a
range query which results in a much smaller dataset to look through. The
latter does a ref which has to scan a much larger dataset, which on
larger sites is significantly slower.