Support tracking of deleted posts
When a post is deleted, the post ID and date is added to an option which is used to augment the existing process of seeking sitemaps to regenerate based on a post's last modified time.

In this instance, we are opting for keeping the post_modified process since it requires less processing as posts are moved—no added filters or functions during that action.