Force $(days|months|years)_being_processed to an array.
Previously, if these options were not set (e.g. when generating a single day's sitemap), it would return the bool `false`, which is not countable.
In PHP versions 7.1-, PHP would proceed with `count` by returning 1. This would result in the logic checks to true unexpectedly, leading in the entire sitemap being rebuilt for every modified post.

In PHP version 7.2, this throws a warning for usage of `count` on a non-countable, so this change also improves 7.2 compatability.