Refactor deleted post tracking to scale better
Previously, there was an array item for each post deleted. Since the purpose of tracking deleted posts is to know which daily sitemaps to regenerate, we only need to keep one item per day. On sites with a large number of posts, the previous behavior could result in an incredibly large option being saved with excessive number of db writes.