Tests: rename test class file
As of PHPUnit 9.1.0, support for test file names which do not match the class name of the test class within, is deprecated. Support is expected to be removed completely in a future PHPUnit version. Effectively, this means that PHPUnit expects PSR4 to be followed for test classes.

While this class is not yet a true TestCase that test classes extend from, it will be, so the naming is taking that into account.