feat: Add XSL stylesheets to XML sitemaps
While XML sitemaps are predominantly for machines to read, they are also easy for humans to navigate to for the purposes of debugging and observing what is being served to the search engine crawlers for site content discovery. We can therefore make it look a little nicer. WordPress core already has this, so we tap into using that existing functionality, adapting it where necessary for  this plugin.

Refactor CoreIntegration out, since we need to re-enable WP core sitemaps to be able to re-use the XSL generation, but then disable the providers so they don’t impact on the content of the XML sitemap generation.

The XSL integration and overrides are in the StylesheetManager.

Includes tests.